Get earthquake information tables from PHIVOLCS earthquake bulletins

Get earthquake information tables from PHIVOLCS earthquake bulletins

eq_get_table(
  .url = "https://earthquake.phivolcs.dost.gov.ph/",
  .year = NULL,
  .month = NULL,
  latest = TRUE
)

Arguments

.url

Base URL for PHIVOLCS earthquake bulletins.

.year

A vector for year (in YYYY format) for which earthquake bulletins are to be retrieved. The earliest year that can be specified is 2018. If set to NULL (default), all years starting from 2018 to present year are used.

.month

A vector for month for which earthquake bulletins are to be retrieved. This can be set as either an integer index (1 for January) or abbreviation (Jan for January) for full name. If set to NULL (default), all months are used.

latest

Logical. Should the latest table of earthquake information be retrieved? Only evaluated if .year = NULL and .month = NULL. If TRUE (default), table of earthquake information for current year and current month is retrieved. Otherwise, all months for all possible years are retrieved.

Value

A list of tibble/s of raw earthquake information from PHIVOLCS.

Examples

eq_get_table()
#> [[1]]
#> # A tibble: 1,148 × 7
#>    date_time_retrieved `Date - Time(Philippine Time)` Latitude\n              …¹
#>    <dttm>              <chr>                          <chr>                     
#>  1 2024-08-14 15:01:41 14 August 2024 - 10:20 PM      13.78                     
#>  2 2024-08-14 15:01:41 14 August 2024 - 09:29 PM      19.25                     
#>  3 2024-08-14 15:01:41 14 August 2024 - 09:06 PM      8.28                      
#>  4 2024-08-14 15:01:41 14 August 2024 - 08:51 PM      5.45                      
#>  5 2024-08-14 15:01:41 14 August 2024 - 08:44 PM      6.33                      
#>  6 2024-08-14 15:01:41 14 August 2024 - 08:32 PM      7.93                      
#>  7 2024-08-14 15:01:41 14 August 2024 - 08:27 PM      21.04                     
#>  8 2024-08-14 15:01:41 14 August 2024 - 08:07 PM      12.99                     
#>  9 2024-08-14 15:01:41 14 August 2024 - 07:55 PM      8.58                      
#> 10 2024-08-14 15:01:41 14 August 2024 - 07:31 PM      4.34                      
#> # ℹ 1,138 more rows
#> # ℹ abbreviated name: ¹​`Latitude\n              (ºN)`
#> # ℹ 4 more variables: `Longitude\n              (ºE)` <chr>,
#> #   `Depth\n              (km)` <chr>, Mag <chr>, Location <chr>
#>