Retrieve earthquake information data from PHIVOLCS bulletins

Retrieve earthquake information data from PHIVOLCS bulletins

Usage

eq_data(.year = NULL, .month = NULL, latest = TRUE, simplify = TRUE)

Arguments

.year

A vector for year (in YYYY format) for which earthquake bulletins are to be retrieved. The earliest year that can be specified is 2018. If set to NULL (default), all years starting from 2018 to present year are used.

.month

A vector for month for which earthquake bulletins are to be retrieved. This can be set as either an integer index (1 for January) or abbreviation (Jan for January) for full name. If set to NULL (default), all months are used.

latest

Logical. Should the latest table of earthquake information be retrieved? Only evaluated if .year = NULL and .month = NULL. If TRUE (default), table of earthquake information for current year and current month is retrieved. Otherwise, all months for all possible years are retrieved.

simplify

Logical. Should output be simplified into a data.frame or tibble? Default is TRUE. Otherwise, a list of processed tibbles of earthquake monitoring data.

Value

A tibble of processed earthquake data.

Examples

eq_data()
#> # A tibble: 1,248 × 7
#>    date_time_retrieved date_time           latitude longitude depth magnitude
#>    <dttm>              <dttm>                 <dbl>     <dbl> <dbl>     <dbl>
#>  1 2024-08-16 06:41:57 2024-08-16 13:48:00     5.02      126.   189       3.9
#>  2 2024-08-16 06:41:57 2024-08-16 13:25:00     7.83      127.    27       2.4
#>  3 2024-08-16 06:41:57 2024-08-16 12:28:00     6.38      124.    19       2.2
#>  4 2024-08-16 06:41:57 2024-08-16 10:40:00    12.8       125.    19       2.6
#>  5 2024-08-16 06:41:57 2024-08-16 10:31:00    10.0       126.    88       4.1
#>  6 2024-08-16 06:41:57 2024-08-16 10:09:00     8.67      126.    30       2.3
#>  7 2024-08-16 06:41:57 2024-08-16 09:53:00    15.0       120.    10       3  
#>  8 2024-08-16 06:41:57 2024-08-16 09:02:00    19.6       121.     1       2.9
#>  9 2024-08-16 06:41:57 2024-08-16 08:20:00    13.2       120.    27       3.7
#> 10 2024-08-16 06:41:57 2024-08-16 07:48:00    12.2       125.    44       1.7
#> # ℹ 1,238 more rows
#> # ℹ 1 more variable: location <chr>